fix(tui): cancel pending Codex login child
